07.05

+++ includes/printf.h
@@ -25,7 +25,7 @@ t_prmim 	ft_parse(const char *format, t_prmim all, va_list args);
 void		ft_print_num(t_prmim all, va_list args);
 
 int			ft_putchar_count(char c, char flag);
-void		putnb_base(int n, int base);
+void		putnb_base(long long n, int base);
 void		putnb_16base(unsigned long n);
 int			digits16_count(unsigned long n);
 void		prekol_s_nulem(t_prmim all);
